feat: 로고 적용
