+++ src/routeTree.gen.ts
@@ -13,11 +13,11 @@ import { createFileRoute } from '@tanstack/react-router'
 // Import Routes
 
 import { Route as rootRoute } from './routes/__root'
+import { Route as IndexImport } from './routes/index'
 
 // Create Virtual Routes
 
 const AboutLazyImport = createFileRoute('/about')()
-const IndexLazyImport = createFileRoute('/')()
 const SignupIndexLazyImport = createFileRoute('/signup/')()
 const MeetingRoomsIndexLazyImport = createFileRoute('/meetingRooms/')()
 const OrganizationsOrganizationIdIndexLazyImport = createFileRoute(
@@ -51,11 +51,11 @@ const AboutLazyRoute = AboutLazyImport.update({
   getParentRoute: () => rootRoute,
 } as any).lazy(() => import('./routes/about.lazy').then((d) => d.Route))
 
-const IndexLazyRoute = IndexLazyImport.update({
+const IndexRoute = IndexImport.update({
   id: '/',
   path: '/',
   getParentRoute: () => rootRoute,
-} as any).lazy(() => import('./routes/index.lazy').then((d) => d.Route))
+} as any)
 
 const SignupIndexLazyRoute = SignupIndexLazyImport.update({
   id: '/signup/',
@@ -164,7 +164,7 @@ declare module '@tanstack/react-router' {
       id: '/'
       path: '/'
       fullPath: '/'
-      preLoaderRoute: typeof IndexLazyImport
+      preLoaderRoute: typeof IndexImport
       parentRoute: typeof rootRoute
     }
     '/about': {
@@ -250,7 +250,7 @@ declare module '@tanstack/react-router' {
 // Create and export the route tree
 
 export interface FileRoutesByFullPath {
-  '/': typeof IndexLazyRoute
+  '/': typeof IndexRoute
   '/about': typeof AboutLazyRoute
   '/meetingRooms': typeof MeetingRoomsIndexLazyRoute
   '/signup': typeof SignupIndexLazyRoute
@@ -265,7 +265,7 @@ export interface FileRoutesByFullPath {
 }
 
 export interface FileRoutesByTo {
-  '/': typeof IndexLazyRoute
+  '/': typeof IndexRoute
   '/about': typeof AboutLazyRoute
   '/meetingRooms': typeof MeetingRoomsIndexLazyRoute
   '/signup': typeof SignupIndexLazyRoute
@@ -281,7 +281,7 @@ export interface FileRoutesByTo {
 
 export interface FileRoutesById {
   __root__: typeof rootRoute
-  '/': typeof IndexLazyRoute
+  '/': typeof IndexRoute
   '/about': typeof AboutLazyRoute
   '/meetingRooms/': typeof MeetingRoomsIndexLazyRoute
   '/signup/': typeof SignupIndexLazyRoute
@@ -342,7 +342,7 @@ export interface FileRouteTypes {
 }
 
 export interface RootRouteChildren {
-  IndexLazyRoute: typeof IndexLazyRoute
+  IndexRoute: typeof IndexRoute
   AboutLazyRoute: typeof AboutLazyRoute
   MeetingRoomsIndexLazyRoute: typeof MeetingRoomsIndexLazyRoute
   SignupIndexLazyRoute: typeof SignupIndexLazyRoute
@@ -357,7 +357,7 @@ export interface RootRouteChildren {
 }
 
 const rootRouteChildren: RootRouteChildren = {
-  IndexLazyRoute: IndexLazyRoute,
+  IndexRoute: IndexRoute,
   AboutLazyRoute: AboutLazyRoute,
   MeetingRoomsIndexLazyRoute: MeetingRoomsIndexLazyRoute,
   SignupIndexLazyRoute: SignupIndexLazyRoute,
@@ -401,7 +401,7 @@ export const routeTree = rootRoute
       ]
     },
     "/": {
-      "filePath": "index.lazy.tsx"
+      "filePath": "index.tsx"
     },
     "/about": {
       "filePath": "about.lazy.tsx"
