+++ src/routes/-components/Footer/Footer.style.ts
@@ -2,20 +2,22 @@ import { Link } from '@tanstack/react-router';
 import styled from 'styled-components';
 
 export const StyledFooter = styled.footer`
-  display: flex;
-  width: 100%;
-  height: 80px;
-  border-top: 1px solid #858585;
-  background: #f9f9f9;
-  align-items: center;
-  justify-content: space-between;
+    display: flex;
+    width: 100%;
+    height: 80px;
+    border-top: 1px solid #858585;
+    background: #f9f9f9;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 90px;
 
-  @media (max-width: 768px) {
-    flex-direction: column;
-    height: auto;
-    padding: 15px 0;
-    gap: 15px;
-  }
+
+    @media (max-width: 768px) {
+        flex-direction: column;
+        height: auto;
+        padding: 15px 0;
+        gap: 15px;
+    }
 `;
 
 export const StyledContainer = styled.div`
@@ -41,21 +43,6 @@ export const StyledLink = styled(Link)`
   @media (max-width: 768px) {
     font-size: 16px;
     margin-left: 0 !important;
-  }
-`;
-
-export const StyledLogo = styled.span`
-  margin-left: 77px;
-  color: #1781ee;
-  font-family: 'Mungyeong Gamhong Apple';
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-
-  @media (max-width: 768px) {
-    margin-left: 0;
-    margin-bottom: 5px;
   }
 `;
